Move to fits tools. Clean up unneeded checks and imports. Change how the date_beg is calculated. Add note to docstring. Add name to pyproject.toml
